Css and HTML changes for integrating HoverNav
